requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
click==8.1.7
fastapi==0.115.3
frozenlist==1.5.0
h11==0.14.0
idna==3.10
loguru==0.7.2
multidict==6.1.0
prometheus_client==0.21.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.16.0
